* Fixes unit test failing because autoIndexCreation was enabled by default * Inline MongoProperties into NhaisMongoClientConfiguration.java - suspect that this MongoProperties bean was being loaded before the config could be overridden by unit tests * Added additional logging to help solve app startup issue
