+++ src/main/java/uk/nhs/digital/nhsconnect/nhais/config/NhaisMongoClientConfiguration.java
@@ -1,10 +1,11 @@
 package uk.nhs.digital.nhsconnect.nhais.config;
 
 
+import com.google.common.base.Strings;
 import com.mongodb.ConnectionString;
 import com.mongodb.MongoClientSettings;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
 
@@ -12,24 +13,72 @@ import org.springframework.data.mongodb.config.AbstractMongoClientConfiguration;
 @Slf4j
 public class NhaisMongoClientConfiguration extends AbstractMongoClientConfiguration {
 
-    @Autowired
-    private MongoProperties mongoProperties;
+    @Value("${nhais.mongodb.database}")
+    private String database;
+
+    @Value("${nhais.mongodb.uri}")
+    private String uri;
+
+    @Value("${nhais.mongodb.host}")
+    private String host;
+
+    @Value("${nhais.mongodb.port}")
+    private String port;
+
+    @Value("${nhais.mongodb.username}")
+    private String username;
+
+    @Value("${nhais.mongodb.password}")
+    private String password;
+
+    @Value("${nhais.mongodb.options}")
+    private String options;
+
+    @Value("${nhais.mongodb.autoIndexCreation}")
+    private String autoIndexCreation;
 
     @Override
     public String getDatabaseName() {
-        return mongoProperties.getDatabase();
+        return this.database;
     }
 
     @Override
     protected void configureClientSettings(MongoClientSettings.Builder builder) {
-        builder.applyConnectionString(new ConnectionString(mongoProperties.createConnectionString()));
+        builder.applyConnectionString(new ConnectionString(this.createConnectionString()));
     }
 
     @Override
     protected boolean autoIndexCreation() {
-        LOGGER.info("Auto index creation is {}", mongoProperties.isAutoIndexCreation());
-        return mongoProperties.isAutoIndexCreation();
+        LOGGER.info("Auto index creation is '{}'", this.autoIndexCreation);
+        return Boolean.parseBoolean(this.autoIndexCreation);
     }
 
+    public String createConnectionString() {
+        if(!Strings.isNullOrEmpty(host)) {
+            LOGGER.info("A value was provided from mongodb host. Generating a connection string from individual properties.");
+            return createConnectionStringFromProperties();
+        } else if(!Strings.isNullOrEmpty(uri)) {
+            LOGGER.info("A mongodb connection string provided in spring.data.mongodb.uri and will be used to configure the database connection.");
+            return uri;
+        } else {
+            LOGGER.error("Mongodb must be configured using a connection string or individual properties. Both uri and host are null or empty");
+            throw new RuntimeException("Missing mongodb connection string and/or properties");
+        }
+    }
 
+    private String createConnectionStringFromProperties() {
+        String cs = "mongodb://";
+        if(!Strings.isNullOrEmpty(username) && !Strings.isNullOrEmpty(password)) {
+            cs += username + ":" + password + "@";
+        } else {
+            LOGGER.info("No mongodb username or password is configured. Will use an anonymous connection.");
+        }
+        cs += host + ":" + port;
+        if(!Strings.isNullOrEmpty(options)) {
+            cs += "?" + options;
+        } else {
+            LOGGER.warn("No options for the mongodb connection string were provided. If connecting to a cluster the driver may not work as expected.");
+        }
+        return cs;
+    }
 }
