+++ src/test/java/uk/nhs/digital/nhsconnect/nhais/IntegrationAdaptorNhaisApplicationTests.java
@@ -1,9 +1,14 @@
 package uk.nhs.digital.nhsconnect.nhais;
 
+import org.junit.Ignore;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.TestPropertySource;
 
 @SpringBootTest
+@TestPropertySource(properties = {
+		"nhais.mongodb.autoIndexCreation=false" // no mongodb instance to create indexes at startup
+})
 class IntegrationAdaptorNhaisApplicationTests {
 
 	@Test
